Add syscall core timer On/Off and some code refactor
