+++ include/exception.h
@@ -21,7 +21,6 @@ void show_invalid_entry_message(int type, unsigned long esr, unsigned long addre
 void sync_exc_router(unsigned long spsr, unsigned long elr, unsigned long esr);
 void sync_svc_handler(unsigned long spsr, unsigned long elr, unsigned long esr);
 void irq_exc_router();
-void print_timestamp();
 void print_pstate_interrupt_mask_bits();
 
 // defined in entry.S.
