+++ include/timer.h
@@ -42,6 +42,8 @@ void dumpyTimerEventList();
 
 void print_timeoutEventInfo(timer_event_t *timeout_event);
 
+void print_timestamp();
+
 uint64_t get_system_time();
 
 // Functions in timer.S
